Move grid measurement helpers together into component
